Fix a connection timeout bug if the lobby is already connected
If the Loby is already connected when the AsyncTask to connect a lobby starts, the lobby will silently drop the request to connect because it's already connected. Thus the AsyncTask is left waiting indefinitely for a callback. It eventually gets timed out after 60s or so at which point it then reports connection _failure_ despite being connected.